refactor(server): split monolithic index.ts into focused modules
Split the 942-line packages/server/src/index.ts into 4 focused modules
to reduce merge conflicts during parallel development:

- stateFilters.ts: toClientState and helper functions for state conversion
- GameServer.ts: Multiplayer game server with player connections
- singlePlayerGame.ts: Single-player game setup with LocalConnection
- index.ts: Clean barrel file re-exporting from modules

This enables parallel work on state filtering, multiplayer logic, and
single-player setup without conflicts in a single large file.
