+++ packages/server/src/index.ts
@@ -1,942 +1,39 @@
 /**
  * @mage-knight/server
- * Server wrapper that connects the game engine to clients
- */
-
-import {
-  type GameState,
-  type Player,
-  type HexState,
-  type HexEnemy,
-  type RuinsToken,
-  type TilePlacement,
-  type TileSlot,
-  type RngState,
-  type TileDeck,
-  type CombatState,
-  type EnemyTokenPiles,
-  type RuinsTokenPiles,
-  createInitialGameState,
-  MageKnightEngine,
-  createEngine,
-  Hero,
-  HEROES,
-  TileId,
-  SiteType,
-  placeTile,
-  hexKey,
-  shuffleWithRng,
-  createEmptyCombatAccumulator,
-  createManaSource,
-  describeEffect,
-  createTileDeck,
-  drawTileFromDeck,
-  generateTileSlots,
-  getValidActions,
-  createEnemyTokenPiles,
-  drawEnemiesForHex,
-  createRuinsTokenPiles,
-  drawRuinsToken,
-  createUnitDecksAndOffer,
-  createSpellDeckAndOffer,
-  createAdvancedActionDeckAndOffer,
-  createArtifactDeck,
-  serializeGameState,
-  deserializeGameState,
-  mineColorToBasicManaColor,
-  countMonasteries,
-  drawMonasteryAdvancedAction,
-} from "@mage-knight/core";
-import type { HexCoord, CardId, HeroId, ScenarioId } from "@mage-knight/shared";
-import { TILE_PLACEMENT_OFFSETS, MAP_SHAPE_CONFIGS, SCENARIO_FIRST_RECONNAISSANCE } from "@mage-knight/shared";
-import {
-  LocalConnection,
-  type GameConnection,
-  type GameEngine,
-  type PlayerAction,
-  type ActionResult,
-  type GameEvent,
-  type ClientGameState,
-  type ClientPlayer,
-  type ClientPlayerUnit,
-  type ClientManaToken,
-  type ClientPendingChoice,
-  type ClientCombatState,
-  type ClientCombatEnemy,
-  type ClientHexEnemy,
-  type ClientRuinsToken,
-  type EventCallback,
-  GAME_PHASE_ROUND,
-  GAME_STARTED,
-  ROUND_PHASE_TACTICS_SELECTION,
-  ALL_DAY_TACTICS,
-  INITIAL_MOVE_POINTS,
-  STARTING_ARMOR,
-  STARTING_COMMAND_TOKENS,
-  STARTING_FAME,
-  STARTING_HAND_LIMIT,
-  STARTING_LEVEL,
-  STARTING_REPUTATION,
-} from "@mage-knight/shared";
-
-// ============================================================================
-// State conversion - full GameState to filtered ClientGameState
-// ============================================================================
-
-/**
- * Convert full GameState to ClientGameState for a specific player.
- * Filters sensitive information (other players' hands, deck contents, etc.)
- */
-export function toClientState(
-  state: GameState,
-  forPlayerId: string
-): ClientGameState {
-  return {
-    phase: state.phase,
-    roundPhase: state.roundPhase,
-    timeOfDay: state.timeOfDay,
-    round: state.round,
-    currentPlayerId: state.turnOrder[state.currentPlayerIndex] ?? "",
-    turnOrder: state.turnOrder,
-    endOfRoundAnnouncedBy: state.endOfRoundAnnouncedBy,
-
-    players: state.players.map((player) =>
-      toClientPlayer(player, forPlayerId)
-    ),
-
-    map: {
-      hexes: Object.fromEntries(
-        Object.entries(state.map.hexes).map(([key, hex]) => [
-          key,
-          {
-            coord: hex.coord,
-            terrain: hex.terrain,
-            tileId: hex.tileId,
-            site: hex.site
-              ? {
-                  type: hex.site.type,
-                  owner: hex.site.owner,
-                  isConquered: hex.site.isConquered,
-                  isBurned: hex.site.isBurned,
-                  ...(hex.site.cityColor && { cityColor: hex.site.cityColor }),
-                  ...(hex.site.mineColor && { mineColor: hex.site.mineColor }),
-                }
-              : null,
-            enemies: hex.enemies.map((enemy): ClientHexEnemy =>
-              toClientHexEnemy(enemy)
-            ),
-            ruinsToken: hex.ruinsToken
-              ? toClientRuinsToken(hex.ruinsToken)
-              : null,
-            shieldTokens: [...hex.shieldTokens],
-            rampagingEnemies: hex.rampagingEnemies.map(String),
-          },
-        ])
-      ),
-      tiles: state.map.tiles.map((tile) => ({
-        centerCoord: tile.centerCoord,
-        revealed: tile.revealed,
-        // Only include tileId for revealed tiles to prevent map hacking
-        ...(tile.revealed && { tileId: tile.tileId }),
-      })),
-      tileSlots: state.map.tileSlots,
-    },
-
-    source: {
-      dice: state.source.dice.map((die) => {
-        // Check if this die is stolen by any player's Mana Steal tactic
-        const isStolenByTactic = state.players.some(
-          (p) => p.tacticState.storedManaDie?.dieId === die.id
-        );
-        return {
-          ...die,
-          isStolenByTactic,
-        };
-      }),
-    },
-
-    offers: {
-      units: state.offers.units,
-      advancedActions: state.offers.advancedActions,
-      spells: state.offers.spells,
-      commonSkills: state.offers.commonSkills,
-      monasteryAdvancedActions: state.offers.monasteryAdvancedActions ?? [],
-    },
-
-    combat: toClientCombatState(state.combat, forPlayerId),
-
-    // Only show deck counts, not contents
-    deckCounts: {
-      spells: state.decks.spells.length,
-      advancedActions: state.decks.advancedActions.length,
-      artifacts: state.decks.artifacts.length,
-      regularUnits: state.decks.regularUnits.length,
-      eliteUnits: state.decks.eliteUnits.length,
-    },
-
-    woundPileCount: state.woundPileCount,
-    scenarioEndTriggered: state.scenarioEndTriggered,
-
-    // Valid actions for this player
-    validActions: getValidActions(state, forPlayerId),
-  };
-}
-
-function toClientPlayer(player: Player, forPlayerId: string): ClientPlayer {
-  const isCurrentPlayer = player.id === forPlayerId;
-
-  return {
-    id: player.id,
-    heroId: player.hero,
-    position: player.position,
-    fame: player.fame,
-    level: player.level,
-    reputation: player.reputation,
-    armor: player.armor,
-    handLimit: player.handLimit,
-    commandTokens: player.commandTokens,
-
-    // Show full hand to self, only count to others
-    hand: isCurrentPlayer ? player.hand : player.hand.length,
-    deckCount: player.deck.length,
-    discardCount: player.discard.length,
-    playArea: player.playArea,
-
-    units: player.units.map(
-      (unit): ClientPlayerUnit => ({
-        instanceId: unit.instanceId,
-        unitId: unit.unitId,
-        state: unit.state,
-        wounded: unit.wounded,
-      })
-    ),
-
-    skills: player.skills,
-    crystals: player.crystals,
-
-    movePoints: player.movePoints,
-    influencePoints: player.influencePoints,
-    pureMana: player.pureMana.map(
-      (token): ClientManaToken => ({
-        color: token.color,
-        source: token.source,
-      })
-    ),
-    hasMovedThisTurn: player.hasMovedThisTurn,
-    hasTakenActionThisTurn: player.hasTakenActionThisTurn,
-    usedManaFromSource: player.usedManaFromSource,
-    playedCardFromHandThisTurn: player.playedCardFromHandThisTurn,
-    isResting: player.isResting,
-
-    knockedOut: player.knockedOut,
-    selectedTacticId: player.selectedTactic,
-    tacticFlipped: player.tacticFlipped,
-
-    pendingChoice: player.pendingChoice
-      ? toClientPendingChoice(player.pendingChoice)
-      : null,
-
-    // Combat accumulator (attack/block values from played cards)
-    combatAccumulator: {
-      attack: {
-        normal: player.combatAccumulator.attack.normal,
-        ranged: player.combatAccumulator.attack.ranged,
-        siege: player.combatAccumulator.attack.siege,
-        normalElements: { ...player.combatAccumulator.attack.normalElements },
-        rangedElements: { ...player.combatAccumulator.attack.rangedElements },
-        siegeElements: { ...player.combatAccumulator.attack.siegeElements },
-      },
-      block: player.combatAccumulator.block,
-      blockElements: { ...player.combatAccumulator.blockElements },
-    },
-
-    // Pending rewards from site conquest
-    pendingRewards: player.pendingRewards,
-
-    // Glade wound choice
-    pendingGladeWoundChoice: player.pendingGladeWoundChoice,
-
-    // Deep mine crystal choice (convert MineColor[] to BasicManaColor[])
-    pendingDeepMineChoice: player.pendingDeepMineChoice
-      ? player.pendingDeepMineChoice.map(mineColorToBasicManaColor)
-      : null,
-
-    // Healing points accumulated this turn
-    healingPoints: player.healingPoints,
-
-    // Stolen mana die from Mana Steal tactic
-    stolenManaDie: player.tacticState.storedManaDie
-      ? {
-          dieId: player.tacticState.storedManaDie.dieId,
-          color: player.tacticState.storedManaDie.color,
-        }
-      : null,
-
-    // Pending level up rewards (skill + AA selection)
-    pendingLevelUpRewards: player.pendingLevelUpRewards,
-
-    // Pending level ups (levels crossed, processed at end of turn)
-    pendingLevelUps: player.pendingLevelUps,
-  };
-}
-
-function toClientPendingChoice(
-  choice: NonNullable<Player["pendingChoice"]>
-): ClientPendingChoice {
-  return {
-    cardId: choice.cardId,
-    skillId: choice.skillId,
-    options: choice.options.map((effect) => ({
-      type: effect.type,
-      description: describeEffect(effect),
-    })),
-  };
-}
-
-/**
- * Convert a HexEnemy to ClientHexEnemy.
- * Masks the token ID when the enemy is unrevealed, only showing the color (token back).
- */
-function toClientHexEnemy(enemy: HexEnemy): ClientHexEnemy {
-  if (enemy.isRevealed) {
-    // Revealed: include the token ID so client can display the enemy face
-    return {
-      color: enemy.color,
-      isRevealed: true,
-      tokenId: String(enemy.tokenId),
-    };
-  } else {
-    // Unrevealed: only show the color (token back), no token ID
-    return {
-      color: enemy.color,
-      isRevealed: false,
-    };
-  }
-}
-
-/**
- * Convert a RuinsToken to ClientRuinsToken.
- * Masks the token ID when unrevealed to prevent cheating - player shouldn't
- * know what's on the token until it's flipped. All unrevealed tokens show
- * the same yellow back.
- */
-function toClientRuinsToken(token: RuinsToken): ClientRuinsToken {
-  if (token.isRevealed) {
-    return {
-      isRevealed: true,
-      tokenId: token.tokenId,
-    };
-  } else {
-    return {
-      isRevealed: false,
-    };
-  }
-}
-
-/**
- * Convert core CombatState to ClientCombatState.
- * Extracts enemy details from definitions for client display.
  *
- * For cooperative assaults, filters enemies to show only those assigned to the
- * requesting player. Each player can only see/target their assigned enemies.
+ * Server wrapper that connects the game engine to clients.
  *
- * @param combat - The full combat state
- * @param forPlayerId - The player requesting the state (used for enemy filtering)
+ * This package provides:
+ * - State filtering (GameState -> ClientGameState for specific players)
+ * - GameServer for multiplayer games
+ * - Single-player game setup with LocalConnection
+ *
+ * ## Module Structure
+ *
+ * | Module | Purpose |
+ * |--------|---------|
+ * | `stateFilters.ts` | Converts GameState to filtered ClientGameState |
+ * | `GameServer.ts` | Multiplayer game server with player connections |
+ * | `singlePlayerGame.ts` | Single-player game setup with LocalConnection |
+ *
+ * @module server
  */
-function toClientCombatState(
-  combat: CombatState | null,
-  forPlayerId: string
-): ClientCombatState | null {
-  if (!combat) return null;
 
-  // Filter enemies for cooperative assaults
-  // If no enemy assignments exist (standard combat), show all enemies
-  const visibleEnemies = combat.enemyAssignments
-    ? combat.enemies.filter((enemy) => {
-        const assignedEnemies = combat.enemyAssignments?.[forPlayerId];
-        return assignedEnemies?.includes(enemy.instanceId) ?? false;
-      })
-    : combat.enemies;
+// State filtering
+export {
+  toClientState,
+  toClientPlayer,
+  toClientPendingChoice,
+  toClientHexEnemy,
+  toClientRuinsToken,
+  toClientCombatState,
+} from "./stateFilters.js";
 
-  return {
-    phase: combat.phase,
-    enemies: visibleEnemies.map(
-      (enemy): ClientCombatEnemy => ({
-        instanceId: enemy.instanceId,
-        enemyId: enemy.enemyId,
-        name: enemy.definition.name,
-        attack: enemy.definition.attack,
-        attackElement: enemy.definition.attackElement,
-        armor: enemy.definition.armor,
-        fame: enemy.definition.fame,
-        abilities: enemy.definition.abilities,
-        resistances: enemy.definition.resistances,
-        isBlocked: enemy.isBlocked,
-        isDefeated: enemy.isDefeated,
-        damageAssigned: enemy.damageAssigned,
-      })
-    ),
-    woundsThisCombat: combat.woundsThisCombat,
-    fameGained: combat.fameGained,
-    isAtFortifiedSite: combat.isAtFortifiedSite,
-  };
-}
-
-// ============================================================================
 // Multiplayer game server
-// ============================================================================
+export { GameServer, createGameServer } from "./GameServer.js";
 
-/**
- * Manages the game and all connected players for multiplayer.
- * Broadcasts events and filtered state to all connected clients.
- */
-export class GameServer {
-  private engine: MageKnightEngine;
-  private state: GameState;
-  private readonly connections: Map<string, EventCallback> = new Map();
-  private readonly seed: number | undefined;
-
-  constructor(seed?: number) {
-    this.engine = createEngine();
-    this.seed = seed;
-    this.state = createInitialGameState(seed);
-  }
-
-  /**
-   * Initialize game with players.
-   * @param playerIds - Array of player IDs (e.g., ["player1", "player2"])
-   * @param heroIds - Optional array of hero IDs corresponding to each player.
-   *                  If not provided, heroes are assigned in default order (Arythea, Tovak, Goldyx, Norowas).
-   * @param scenarioId - Optional scenario to play. Defaults to First Reconnaissance.
-   */
-  initializeGame(
-    playerIds: readonly string[],
-    heroIds?: readonly HeroId[],
-    scenarioId: ScenarioId = SCENARIO_FIRST_RECONNAISSANCE
-  ): void {
-    this.state = this.createGameWithPlayers(playerIds, heroIds, scenarioId);
-
-    // Broadcast initial state to all connected players
-    this.broadcastState([
-      {
-        type: GAME_STARTED,
-        playerCount: playerIds.length,
-        scenario: scenarioId,
-      },
-    ]);
-  }
-
-  /**
-   * Player connects and provides their callback for receiving events/state.
-   */
-  connect(playerId: string, callback: EventCallback): void {
-    this.connections.set(playerId, callback);
-
-    // Send current state to newly connected player
-    const clientState = toClientState(this.state, playerId);
-    callback([], clientState);
-  }
-
-  /**
-   * Player disconnects.
-   */
-  disconnect(playerId: string): void {
-    this.connections.delete(playerId);
-  }
-
-  /**
-   * Player sends an action. Broadcasts result to ALL connected players.
-   */
-  handleAction(playerId: string, action: PlayerAction): void {
-    // Process through engine
-    const result = this.engine.processAction(this.state, playerId, action);
-
-    // Update server state
-    this.state = result.state;
-
-    // Broadcast to all connected players
-    this.broadcastState(result.events);
-  }
-
-  /**
-   * Get current state (for testing/debugging).
-   */
-  getState(): GameState {
-    return this.state;
-  }
-
-  /**
-   * Get filtered state for a specific player.
-   */
-  getStateForPlayer(playerId: string): ClientGameState {
-    return toClientState(this.state, playerId);
-  }
-
-  /**
-   * Serialize current game state for saving.
-   */
-  saveGame(): string {
-    return serializeGameState(this.state);
-  }
-
-  /**
-   * Load a saved game state, replacing current state.
-   * Broadcasts updated state to all connected clients.
-   */
-  loadGame(json: string): void {
-    this.state = deserializeGameState(json);
-    this.broadcastState([]);
-  }
-
-  /**
-   * Broadcast events and filtered state to all connections.
-   */
-  private broadcastState(events: readonly GameEvent[]): void {
-    for (const [playerId, callback] of this.connections) {
-      const clientState = toClientState(this.state, playerId);
-      callback(events, clientState);
-    }
-  }
-
-  /**
-   * Create initial game state with players.
-   * Places starting tile + 2 initial countryside tiles and positions all players on the portal hex.
-   * Uses seeded RNG for reproducible initial deck shuffles.
-   * @param playerIds - Array of player IDs
-   * @param heroIds - Optional array of hero IDs for each player. Falls back to default order.
-   * @param scenarioId - Scenario to play
-   */
-  private createGameWithPlayers(
-    playerIds: readonly string[],
-    heroIds?: readonly HeroId[],
-    scenarioId: ScenarioId = SCENARIO_FIRST_RECONNAISSANCE
-  ): GameState {
-    const baseState = createInitialGameState(this.seed, scenarioId);
-
-    // Initialize tile deck based on scenario configuration
-    const { tileDeck: initialDeck, rng: rngAfterDeck } = createTileDeck(
-      baseState.scenarioConfig,
-      baseState.rng
-    );
-
-    // Initialize enemy token piles FIRST so we can draw enemies as tiles are placed
-    const { piles: initialEnemyPiles, rng: rngAfterEnemyInit } = createEnemyTokenPiles(rngAfterDeck);
-    let currentEnemyPiles: EnemyTokenPiles = initialEnemyPiles;
-
-    // Initialize ruins token piles for Ancient Ruins yellow tokens
-    const { piles: initialRuinsPiles, rng: rngAfterRuinsInit } = createRuinsTokenPiles(rngAfterEnemyInit);
-    let currentRuinsPiles: RuinsTokenPiles = initialRuinsPiles;
-    let currentRng: RngState = rngAfterRuinsInit;
-
-    // Calculate total tiles for scenario
-    const totalTiles =
-      1 + // starting tile
-      baseState.scenarioConfig.countrysideTileCount +
-      baseState.scenarioConfig.coreTileCount +
-      baseState.scenarioConfig.cityTileCount;
-
-    // Generate tile slots based on map shape
-    const slotsMap = generateTileSlots(
-      baseState.scenarioConfig.mapShape,
-      totalTiles
-    );
-
-    // Get map shape configuration
-    const mapShapeConfig = MAP_SHAPE_CONFIGS[baseState.scenarioConfig.mapShape];
-
-    // Convert to Record and mark starting position as filled
-    const tileSlots: Record<string, TileSlot> = {};
-    for (const [key, slot] of slotsMap) {
-      tileSlots[key] = slot;
-    }
-
-    // Place starting tile at origin (tile type from map shape config)
-    const tileOrigin: HexCoord = { q: 0, r: 0 };
-    const startingTileId = mapShapeConfig.startingTile === "starting_a"
-      ? TileId.StartingTileA
-      : TileId.StartingTileB;
-    const startingTileHexes = placeTile(startingTileId, tileOrigin);
-
-    // Mark starting slot as filled
-    const originKey = hexKey(tileOrigin);
-    if (tileSlots[originKey]) {
-      tileSlots[originKey] = { ...tileSlots[originKey], filled: true };
-    }
-
-    // Build hex map starting with the starting tile (no enemies on starting tile)
-    const hexes: Record<string, HexState> = {};
-    for (const hex of startingTileHexes) {
-      const key = hexKey(hex.coord);
-      hexes[key] = hex;
-    }
-
-    // Track tile placements
-    const tiles: TilePlacement[] = [
-      {
-        tileId: startingTileId,
-        centerCoord: tileOrigin,
-        revealed: true,
-      },
-    ];
-
-    // Place initial countryside tiles adjacent to the starting tile
-    // Number and positions determined by map shape configuration
-    const initialTilePositions: HexCoord[] = mapShapeConfig.initialTilePositions.map(
-      (dir) => TILE_PLACEMENT_OFFSETS[dir]
-    );
-
-    // Track all hexes from initial tiles to count monasteries later
-    const initialTileHexes: HexState[] = [];
-
-    let currentDeck: TileDeck = initialDeck;
-    for (const position of initialTilePositions) {
-      const drawResult = drawTileFromDeck(currentDeck);
-      if (drawResult) {
-        const { tileId, updatedDeck } = drawResult;
-        currentDeck = updatedDeck;
-
-        // Place the tile and add its hexes to the map
-        const tileHexes = placeTile(tileId, position);
-        initialTileHexes.push(...tileHexes);
-        for (const hex of tileHexes) {
-          const key = hexKey(hex.coord);
-
-          // Draw enemies for hexes that need them (sites with defenders, rampaging enemies)
-          const rampagingTypes = [...hex.rampagingEnemies];
-
-          const siteType = hex.site?.type ?? null;
-
-          const { enemies, piles: enemyPiles, rng: rngAfterEnemies } = drawEnemiesForHex(
-            rampagingTypes,
-            siteType,
-            currentEnemyPiles,
-            currentRng,
-            baseState.timeOfDay
-          );
-
-          currentEnemyPiles = enemyPiles;
-          currentRng = rngAfterEnemies;
-
-          // Draw ruins token for Ancient Ruins sites
-          let ruinsToken = hex.ruinsToken;
-          if (siteType === SiteType.AncientRuins) {
-            const ruinsResult = drawRuinsToken(
-              currentRuinsPiles,
-              currentRng,
-              baseState.timeOfDay
-            );
-            ruinsToken = ruinsResult.token;
-            currentRuinsPiles = ruinsResult.piles;
-            currentRng = ruinsResult.rng;
-          }
-
-          // Update hex with drawn enemies and ruins token
-          // NOTE: Preserve rampagingEnemies marker - it's needed for movement validation
-          // (validateNotBlockedByRampaging checks BOTH rampagingEnemies.length > 0 AND enemies.length > 0)
-          hexes[key] = {
-            ...hex,
-            enemies: enemies,
-            ruinsToken,
-          };
-        }
-
-        tiles.push({
-          tileId,
-          centerCoord: position,
-          revealed: true,
-        });
-
-        // Mark slot as filled
-        const posKey = hexKey(position);
-        if (tileSlots[posKey]) {
-          tileSlots[posKey] = { ...tileSlots[posKey], filled: true };
-        }
-      }
-    }
-
-    // Find portal hex for player starting position
-    const portalHex = startingTileHexes.find((h) => h.site?.type === SiteType.Portal);
-    const startPosition = portalHex?.coord ?? { q: 0, r: 0 };
-
-    // Create players on the portal with seeded RNG for deck shuffles
-    const players: Player[] = [];
-
-    for (let index = 0; index < playerIds.length; index++) {
-      const id = playerIds[index];
-      const heroId = heroIds?.[index];
-      if (id !== undefined) {
-        const { player, rng } = this.createPlayer(id, index, heroId, startPosition, currentRng);
-        players.push(player);
-        currentRng = rng;
-      }
-    }
-
-    // Set up tactics selection phase
-    // In solo play, the single player selects first
-    // In multiplayer, selection order is reverse fame (lowest fame picks first)
-    // Since all players start at 0 fame, use player order
-    const tacticsSelectionOrder = [...playerIds];
-
-    // Initialize mana source with dice (playerCount + 2 dice)
-    const { source, rng: rngAfterMana } = createManaSource(
-      playerIds.length,
-      baseState.timeOfDay,
-      currentRng
-    );
-
-    // Initialize unit decks and populate initial unit offer
-    const {
-      decks: unitDecks,
-      unitOffer,
-      rng: rngAfterUnits,
-    } = createUnitDecksAndOffer(
-      baseState.scenarioConfig,
-      playerIds.length,
-      rngAfterMana
-    );
-
-    // Initialize spell deck and populate initial spell offer
-    const {
-      spellDeck,
-      spellOffer,
-      rng: rngAfterSpells,
-    } = createSpellDeckAndOffer(rngAfterUnits);
-
-    // Initialize advanced action deck and populate initial AA offer
-    const {
-      advancedActionDeck,
-      advancedActionOffer,
-      rng: rngAfterAA,
-    } = createAdvancedActionDeckAndOffer(rngAfterSpells);
-
-    // Initialize artifact deck
-    const {
-      artifactDeck,
-      rng: rngAfterArtifacts,
-    } = createArtifactDeck(rngAfterAA);
-
-    // Draw Advanced Actions for any monasteries on initial tiles
-    const monasteryCount = countMonasteries(initialTileHexes);
-    let currentAADeck = advancedActionDeck;
-    let monasteryAAs: readonly CardId[] = [];
-    const rngAfterMonasteries = rngAfterArtifacts;
-
-    for (let i = 0; i < monasteryCount; i++) {
-      const result = drawMonasteryAdvancedAction(
-        { ...baseState.offers, monasteryAdvancedActions: monasteryAAs },
-        { ...baseState.decks, advancedActions: currentAADeck }
-      );
-      currentAADeck = result.decks.advancedActions;
-      monasteryAAs = result.offers.monasteryAdvancedActions;
-    }
-
-    return {
-      ...baseState,
-      phase: GAME_PHASE_ROUND,
-      roundPhase: ROUND_PHASE_TACTICS_SELECTION,
-      availableTactics: [...ALL_DAY_TACTICS],
-      tacticsSelectionOrder,
-      currentTacticSelector: tacticsSelectionOrder[0] ?? null,
-      turnOrder: playerIds,
-      currentPlayerIndex: 0,
-      players,
-      source,
-      enemyTokens: currentEnemyPiles, // Enemy piles after drawing for initial tiles
-      ruinsTokens: currentRuinsPiles, // Ruins piles after drawing for initial tiles
-      rng: rngAfterMonasteries, // Updated RNG state after all shuffles
-      decks: {
-        ...baseState.decks,
-        regularUnits: unitDecks.regularUnits,
-        eliteUnits: unitDecks.eliteUnits,
-        spells: spellDeck,
-        advancedActions: currentAADeck,
-        artifacts: artifactDeck,
-      },
-      offers: {
-        ...baseState.offers,
-        units: unitOffer,
-        spells: { cards: [...spellOffer] },
-        advancedActions: { cards: [...advancedActionOffer] },
-        monasteryAdvancedActions: monasteryAAs,
-      },
-      map: {
-        ...baseState.map,
-        hexes,
-        tiles,
-        tileDeck: currentDeck, // Remaining tiles after initial placement
-        tileSlots, // Tile slot grid for map shape constraints
-      },
-    };
-  }
-
-  /**
-   * Create a player with default values.
-   * Shuffles the hero's starting deck using seeded RNG and draws an initial hand.
-   * Returns both the player and the updated RNG state.
-   * @param id - Player ID
-   * @param index - Player index (0-based)
-   * @param heroId - Optional hero ID. If not provided, falls back to default order.
-   * @param position - Starting position on the map
-   * @param rng - Current RNG state
-   */
-  private createPlayer(
-    id: string,
-    index: number,
-    heroId: HeroId | undefined,
-    position: { q: number; r: number },
-    rng: RngState
-  ): { player: Player; rng: RngState } {
-    // Default hero assignment (fallback for legacy/test code)
-    const defaultHeroes: readonly Hero[] = [
-      Hero.Arythea,
-      Hero.Tovak,
-      Hero.Goldyx,
-      Hero.Norowas,
-    ];
-
-    // Use provided heroId, or fall back to default by index
-    const hero = heroId
-      ? (heroId as Hero) // HeroId strings match Hero enum values
-      : (defaultHeroes[index % defaultHeroes.length] ?? Hero.Arythea);
-    const heroDefinition = HEROES[hero];
-
-    // Create and shuffle starting deck with seeded RNG
-    const allCards = [...heroDefinition.startingCards];
-    const { result: shuffledDeck, rng: newRng } = shuffleWithRng(allCards, rng);
-
-    // Draw starting hand
-    const handLimit = STARTING_HAND_LIMIT;
-    const startingHand = shuffledDeck.slice(0, handLimit);
-    const remainingDeck = shuffledDeck.slice(handLimit);
-
-    const player: Player = {
-      id,
-      hero,
-      position, // Start on the portal
-      fame: STARTING_FAME,
-      level: STARTING_LEVEL,
-      reputation: STARTING_REPUTATION,
-      armor: STARTING_ARMOR,
-      handLimit,
-      commandTokens: STARTING_COMMAND_TOKENS,
-      hand: startingHand,
-      deck: remainingDeck,
-      discard: [],
-      units: [],
-      skills: [],
-      skillCooldowns: {
-        usedThisRound: [],
-        usedThisTurn: [],
-        usedThisCombat: [],
-        activeUntilNextTurn: [],
-      },
-      skillFlipState: {
-        flippedSkills: [],
-      },
-      crystals: { red: 0, blue: 0, green: 0, white: 0 },
-      selectedTactic: null,
-      tacticFlipped: false,
-      tacticState: {},
-      pendingTacticDecision: null,
-      beforeTurnTacticPending: false,
-      knockedOut: false,
-      movePoints: INITIAL_MOVE_POINTS,
-      influencePoints: 0,
-      playArea: [],
-      pureMana: [],
-      usedManaFromSource: false,
-      usedDieIds: [],
-      manaDrawDieIds: [],
-      hasMovedThisTurn: false,
-      hasTakenActionThisTurn: false,
-      hasCombattedThisTurn: false,
-      playedCardFromHandThisTurn: false,
-      hasPlunderedThisTurn: false,
-      hasRecruitedUnitThisTurn: false,
-      manaUsedThisTurn: [],
-      combatAccumulator: createEmptyCombatAccumulator(),
-      pendingChoice: null,
-      pendingLevelUps: [],
-      pendingLevelUpRewards: [],
-      remainingHeroSkills: [...heroDefinition.skills],
-      pendingRewards: [],
-      pendingGladeWoundChoice: false,
-      pendingDeepMineChoice: null,
-      healingPoints: 0,
-      removedCards: [],
-      isResting: false,
-      roundOrderTokenFlipped: false,
-    };
-
-    return { player, rng: newRng };
-  }
-}
-
-// ============================================================================
-// Single-player game instance (uses LocalConnection)
-// ============================================================================
-
-export interface GameInstance {
-  readonly engine: GameEngine;
-  readonly connection: GameConnection;
-}
-
-/**
- * Adapter to make MageKnightEngine compatible with GameEngine interface.
- * Used for LocalConnection which expects GameEngine.
- */
-class EngineAdapter implements GameEngine {
-  private state: GameState;
-  private readonly mageKnightEngine: MageKnightEngine;
-
-  constructor(
-    private readonly playerId: string,
-    seed?: number
-  ) {
-    this.mageKnightEngine = createEngine();
-    this.state = createInitialGameState(seed);
-  }
-
-  processAction(playerId: string, action: PlayerAction): ActionResult {
-    const result = this.mageKnightEngine.processAction(
-      this.state,
-      playerId,
-      action
-    );
-    this.state = result.state;
-    return {
-      events: result.events,
-      state: toClientState(result.state, this.playerId),
-    };
-  }
-}
-
-// ============================================================================
-// Factory functions
-// ============================================================================
-
-/**
- * Create a single-player game instance with LocalConnection.
- * @param playerId - The player's ID
- * @param seed - Optional seed for reproducible RNG
- */
-export function createGame(playerId: string, seed?: number): GameInstance {
-  const engineAdapter = new EngineAdapter(playerId, seed);
-  const connection = new LocalConnection(engineAdapter, playerId);
-
-  return {
-    engine: engineAdapter,
-    connection,
-  };
-}
-
-/**
- * Create a multiplayer game server.
- * @param seed - Optional seed for reproducible RNG
- */
-export function createGameServer(seed?: number): GameServer {
-  return new GameServer(seed);
-}
+// Single-player game
+export { createGame, type GameInstance } from "./singlePlayerGame.js";
 
 // Re-export types
 export type { EventCallback } from "@mage-knight/shared";
